pkg/generated/models/firewall_sequence_test.go

package models

import "testing"


import "fmt"
func TestFirewallSequence(t *testing.T) {
    model := MakeFirewallSequence()
    fmt.Println(model)
}


func TestFirewallSequenceFQName(t *testing.T) {
    obj := MakeFirewallSequence()
    if fqname := obj.GetFQName(); len(fqname != 0) {
        t.Errorf("Initial FQName is not empty for FirewallSequence")
    }
    parent := "fake-parent"
    firstName := "first name"
    fakeFQName := []string{"some", "fake", firstName}
    newname := "fake-me"
    obj.SetFQName(parent, fakeFQName)
    fqname = obj.GetFQName()
    if len(fqname) != len(fakeFQName) {
        t.Errorf("Wrong FQName length, should be %v instead of %v", len(fakeFQName), len(fqname))
    }
    if fqname[len(fqname)-1] != newname {
        t.Errorf("Name should be %v after SetFQName", newname)
    }
}

